requirements.txt
streamlit==1.28.0
scikit-learn==1.3.0
joblib==1.3.2
python-docx==1.1.0
pdfplumber==0.10.3
pandas==2.0.3